requirements.txt
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Python-dotenv==1.0.0
flask==3.0.0
Flask-SQLAlchemy==3.1.1
Gunicorn==21.2.0

